Code deduplication.
